moved task storage to separate file (refactoring)
